refactor: replace SocketActorKey with RawHandle for improved handle management
- Updated the driver and related components to replace instances of `SocketActorKey` with `RawHandle`, enhancing clarity and type safety in handle management.
- Refactored method signatures and type aliases to align with the new handle structure, ensuring consistent access to raw file descriptors.
- Improved socket generation logic by utilizing thread-local storage for better performance and resource management.
- Cleaned up the codebase by removing unused methods and enhancing error handling processes, reinforcing overall robustness and maintainability.

+++ platform-drivers/veloq-driver-iocp/src/rio/core/registry.rs
@@ -219,7 +219,7 @@ impl RioRegistry {
 
         env.dispatch
             .create_rq(RioRqConfig {
-                socket: handle.raw().as_socket() as usize,
+                socket: handle.raw().as_socket(),
                 max_outstanding_recvs,
                 max_receive_data_buffers: 1,
                 max_outstanding_sends,
